Make Hct.toInt() public
The `toInt()` method of `Hct`, which returns the RGBA representation only has package visibility, where as it's other properties (e.g `hue`, `chroma`, and `tone`) are `public`. This means that when consuming this repo as a SPM dependency it's possible to create a `Hct` value from an RGBA value and modify it, but it's not possible to convert it back to an RGBA value.